api/pkg/features: move cloud under “enterprise”
--- sturdy ---
user_id: 847dfd0c-49bf-40c5-8870-74a12fca0d60
view_id: 128badfe-5c58-487e-adfd-c4f1e4fd160f
workspace_id: e558c894-4674-415e-9632-186b41af1e09
