+++ api/pkg/features/graphql/module_enterprise.go
@@ -5,9 +5,9 @@ package graphql
 
 import (
 	"getsturdy.com/api/pkg/di"
-	"getsturdy.com/api/pkg/features/graphql/enterprise"
+	"getsturdy.com/api/pkg/features/graphql/enterprise/selfhosted"
 )
 
 func Module(c *di.Container) {
-	c.Register(enterprise.NewFeaturesRootResolver)
+	c.Register(selfhosted.NewFeaturesRootResolver)
 }
